// src/types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  Decimal: any;
};

export type Asset = {
  __typename?: 'Asset';
  fireblocksTicker?: Maybe<Scalars['String']>;
  id?: Maybe<Scalars['Int']>;
  name: Scalars['String'];
  ticker?: Maybe<Scalars['String']>;
  type: Scalars['String'];
};

export type AssetBalanceType = {
  __typename?: 'AssetBalanceType';
  balanceTypeName?: Maybe<Scalars['String']>;
};

export type CollateralSummary = {
  __typename?: 'CollateralSummary';
  assets?: Maybe<Array<Maybe<WalletAsset>>>;
  totalBalance?: Maybe<Scalars['String']>;
  totalWallets?: Maybe<Scalars['Int']>;
};

export type CollateralWallet = {
  __typename?: 'CollateralWallet';
  assets?: Maybe<Array<Maybe<WalletAsset>>>;
  loanEndDate?: Maybe<Scalars['DateTime']>;
  ltvFeePercentage?: Maybe<Scalars['Decimal']>;
  minimumPayment?: Maybe<Scalars['Decimal']>;
  minimumPaymentDueDate?: Maybe<Scalars['DateTime']>;
  nickname?: Maybe<Scalars['String']>;
  outstandingPaymentBalance?: Maybe<Scalars['Decimal']>;
  paidOnLoanAmount?: Maybe<Scalars['Decimal']>;
  totalBalance?: Maybe<Scalars['Decimal']>;
  totalLoanAmount?: Maybe<Scalars['Decimal']>;
  type?: Maybe<Scalars['String']>;
  walletId?: Maybe<Scalars['String']>;
};
// ...
export type StakingWallet = {
  __typename?: 'StakingWallet';
  assets?: Maybe<Array<Maybe<WalletAsset>>>;
  availableRewards?: Maybe<Scalars['Decimal']>;
  balance?: Maybe<Scalars['Decimal']>;
  nickname?: Maybe<Scalars['String']>;
  paidRewards?: Maybe<Scalars['Decimal']>;
  pendingRewards?: Maybe<Scalars['String']>;
  stakedBalance?: Maybe<Scalars['Decimal']>;
  totalBalance?: Maybe<Scalars['Decimal']>;
  type?: Maybe<Scalars['String']>;
  walletId?: Maybe<Scalars['String']>;
};
// ...
export type TransactionalWallet = {
  __typename?: 'TransactionalWallet';
  assets?: Maybe<Array<Maybe<WalletAsset>>>;
  balance?: Maybe<Scalars['Decimal']>;
  nickname?: Maybe<Scalars['String']>;
  pendingRewards?: Maybe<Scalars['String']>;
  totalBalance?: Maybe<Scalars['Decimal']>;
  type?: Maybe<Scalars['String']>;
  walletId?: Maybe<Scalars['String']>;
};
// ...
export type WalletAsset = {
  __typename?: 'walletAsset';
  assetBalanceType?: Maybe<AssetBalanceType>;
  assets?: Maybe<Asset>;
  balance?: Maybe<Scalars['Decimal']>;
  price?: Maybe<Scalars['String']>;
  sevenDayPercChange?: Maybe<Scalars['Decimal']>;
  usdPrice?: Maybe<Scalars['Decimal']>;
  walletAssetsID: Scalars['String'];
};